refactor: inputBar component by extracting form logic to separate Form component
